The installation script was improved.

+++ administration_tools/mk-dist.sh
@@ -13,10 +13,19 @@ if [ "$1" = "-h" ] || [ "$1" = "--help" ]; then
     echo "OPTIONS may be:"
     echo
     echo "   -h --help: This help"
+    echo "   --keep : Do not delete distribution directory after"
+    echo "            the tar file is created."
     exit 0
 fi
 
-cd "$MYDIR/.."
+if [ "$1" == "--keep" ]; then
+    keep="yes"
+fi
+
+if ! cd "$MYDIR/.."; then
+    echo "Error, cd $MYDIR/.. failed" >&2
+    exit 1
+fi
 
 if [ ! -d "doc/_build" ]; then
     echo "Run ./doc-rebuild.sh first">&2
@@ -40,8 +49,11 @@ bin/git-gq doc | rst2man > "$DISTPATH"/man/man1/git-gq.1
 mkdir -p "$DISTPATH"/profile.d
 bin/git-gq bashcompletion > "$DISTPATH"/profile.d/git-gq.sh
 cp "$MYDIR/install.sh" "$DISTPATH"
+cp "$MYDIR/git-gq-uninstall.sh" "$DISTPATH"
 cp -a doc/_build/html "$DISTPATH/doc"
 
 (cd dist && tar -czf "$DISTDIR.tar.gz" "$DISTDIR")
 
-rm -rf "$DISTPATH"
+if [ -z "$keep" ]; then
+    rm -rf "$DISTPATH"
+fi
